+++ modules/channel/src/main/scala/higherkindness/mu/rpc/protocol/service.scala
@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-package higherkindness.mu.rpc.healthcheck
+package higherkindness.mu.rpc.protocol
 
-import cats.kernel.Order
-import cats.instances.string._
-import higherkindness.mu.rpc.healthcheck.unary.handler.HealthCheck
+import higherkindness.mu.rpc.internal.serviceImpl
 
-object ordering {
-  implicit val orderForHealthCheck: Order[HealthCheck] = new HealthCheckOrder
+import scala.annotation.{compileTimeOnly, StaticAnnotation}
 
-  class HealthCheckOrder(implicit O: Order[String]) extends Order[HealthCheck] {
-
-    override def eqv(x: HealthCheck, y: HealthCheck): Boolean =
-      O.eqv(x.nameService, y.nameService)
-
-    def compare(x: HealthCheck, y: HealthCheck): Int =
-      O.compare(x.nameService, y.nameService)
-  }
+// $COVERAGE-OFF$
+@compileTimeOnly("enable macro paradise to expand @service macro annotations")
+class service(
+    val serializationType: SerializationType,
+    val compressionType: CompressionType = Identity,
+    val namespace: Option[String] = None,
+    val methodNameStyle: MethodNameStyle = Unchanged
+) extends StaticAnnotation {
+  def macroTransform(annottees: Any*): Any = macro serviceImpl.service
 }
+// $COVERAGE-ON$
