+++ modules/config/src/main/scala/higherkindness/mu/rpc/config.scala
@@ -22,10 +22,6 @@ import com.typesafe.config.{Config, ConfigFactory}
 
 package object config {
 
-  trait ConfigM[F[_]] {
-    def load: F[Config]
-  }
-
   implicit def syncConfigM[F[_]](implicit F: Sync[F]): ConfigM[F] = new ConfigM[F] {
     def load: F[Config] = F.delay(ConfigSource.fromConfig(ConfigFactory.load()).loadOrThrow[Config])
   }
